fix not checking that a declared owner in a transaction actually owns the token
the code was previously not directly checking the owners, but checking the balance owned. If balance was not enough to transfer, the transfer was rejected
We deleted the balance checking code, as there's now a fixed amount of 1 and a simplified ledger map, but without adding back an ownership check.