Add  transfer_collection bulk transfer entrypoint
This is used by the marketplace contract to swap a whole collection to the marketplace after minting

Calling the entrypoint only works on a fresh collection, ie. a collection that is still only referenced in the collection ledger (entirely in lazy ledger mode), and which no token has been written to the traditional token ledger

This limitation is to avoid corner cases and simplify the model: you can either transfer a whole, fresh collection (cheap),  or single tokens (traditional cost).

Managing the mixed cases would create not fully predictable results for the caller (the collection is transferred but token already transferred would not (as the contract gives the priority to the ledger truth over the collection ledger), and would need the contract to handle the case where the contract tokens might be back in the collection owner 's ownership after a token transfer to another owner (written in the ledger), and thus part of the collection from their point of view, but not from the point of view of the contract